03_litellm_openai_agent/pyproject.toml
[project]
name = "03-litellm-openai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "litellm>=1.72.0",
    "python-dotenv>=1.1.0",
]
